KL-138/rename: rename, move files
- 폴더 위치 이동 및 모호한 이름 변경
