+++ src/components/Navbar/SearchModal.jsx
@@ -4,7 +4,7 @@ import { Modal, ConfigProvider } from 'antd'
 import { useShallow } from 'zustand/react/shallow'
 import { debounce } from 'lodash-es'
 
-import Icons from '../Common/Icons'
+import Icons from '../Icons/Icons'
 import theme from '../../styles/theme'
 import { useModalStore, useCurrentPageStore } from './stores/navbarStores'
