+++ src/index.js
@@ -1,12 +1,12 @@
-import React from 'react'
+import React, { Suspense, lazy } from 'react'
 import ReactDOM from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import PageLayout from './PageLayout'
-import ErrorPage from './pages/ErrorPage'
-import Home from './components/Home/Home'
-import ReviewDetailPage from './components/Review/ReviewDetailPage'
-import SubmitPage from './components/SubmitPage/SubmitPage'
-import FeedPage from './pages/FeedPage/FeedPage'
+
+const HomePage = lazy(() => import('./pages/HomePage/HomePage'))
+const ReviewPage = lazy(() => import('./pages/ReviewPage/ReviewPage'))
+const SubmitPage = lazy(() => import('./pages/SubmitPage/SubmitPage'))
+const ErrorPage = lazy(() => import('./pages/ErrorPage'))
 
 const router = createBrowserRouter([
   {
@@ -15,19 +15,27 @@ const router = createBrowserRouter([
     children: [
       {
         path: '',
-        element: <Home />,
+        element: (
+          <Suspense fallback={<div>Loading...</div>}>
+            <HomePage />
+          </Suspense>
+        ),
       },
       {
         path: 'review',
-        element: <ReviewDetailPage />,
+        element: (
+          <Suspense fallback={<div>Loading...</div>}>
+            <ReviewPage />
+          </Suspense>
+        ),
       },
       {
         path: 'submit',
-        element: <SubmitPage />,
-      },
-      {
-        path: 'feed',
-        element: <FeedPage />,
+        element: (
+          <Suspense fallback={<div>Loading...</div>}>
+            <SubmitPage />
+          </Suspense>
+        ),
       },
     ],
   },
@@ -39,8 +47,3 @@ const router = createBrowserRouter([
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
 root.render(<RouterProvider router={router} />)
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals()
