+++ src/pages/FeedPage/SelectedField/CategoryField.jsx
@@ -3,7 +3,10 @@ import PropTypes from 'prop-types'
 import { CloseOutlined } from '@ant-design/icons'
 import useSelectedCategory from '../../../hooks/useSelectedCategory'
 import useSelectedSubCategory from '../../../hooks/useSelectedSubCategory'
-import { BlueFieldTag, WhiteFieldTag } from '../../../components/tag/Tags.style'
+import {
+  BlueFieldTag,
+  WhiteFieldTag,
+} from '../../../components/Tags/Tags.style'
 
 function SubCategoryField({ categoryId }) {
   const { selectedSubCategory, setSelectedSubCategory } =
